+++ Bds/models/CreateBdsInstanceDetails.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2021, Oracle and/or its affiliates. All rights reserved.
  * This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
  */
 
@@ -92,9 +92,6 @@ namespace Oci.BdsService.Models
         [JsonProperty(PropertyName = "isSecure")]
         public System.Nullable<bool> IsSecure { get; set; }
         
-        /// <value>
-        /// Additional configuration of customer's network.
-        /// </value>
         [JsonProperty(PropertyName = "networkConfig")]
         public NetworkConfig NetworkConfig { get; set; }