+++ Bds/models/OperationTypes.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2021, Oracle and/or its affiliates. All rights reserved.
  * This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
  */
 
@@ -37,6 +37,10 @@ namespace Oci.BdsService.Models
       [EnumMember(Value = "UPDATE_INFRA")]
       UpdateInfra,
       [EnumMember(Value = "RESTART_NODE")]
-      RestartNode
+      RestartNode,
+      [EnumMember(Value = "AUTOSCALE_CONFIG")]
+      AutoscaleConfig,
+      [EnumMember(Value = "AUTOSCALE_RUN")]
+      AutoscaleRun
   }
 }