+++ Bds/requests/RemoveCloudSqlRequest.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2021, Oracle and/or its affiliates. All rights reserved.
  * This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
  */
 
@@ -55,5 +55,16 @@ namespace Oci.BdsService.Requests
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "if-match")]
         public string IfMatch { get; set; }
+        
+        /// <value>
+        /// A token that uniquely identifies a request so it can be retried in case of a timeout or
+        /// server error without risk of executing that same action again. Retry tokens expire after 24
+        /// hours, but can be invalidated before then due to conflicting operations. For example, if a resource
+        /// has been deleted and purged from the system, then a retry of the original creation request
+        /// might be rejected.
+        /// 
+        /// </value>
+        [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "opc-retry-token")]
+        public string OpcRetryToken { get; set; }
     }
 }