+++ Objectstoragetests/Transfer/ProgressTrackingInputStreamTests.cs
@@ -1,4 +1,9 @@
-﻿using System;
+/*
+ * Copyright (c) 2020, 2021, Oracle and/or its affiliates. All rights reserved.
+ * This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
+ */
+
+using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
@@ -7,7 +12,7 @@ using Oci.Common;
 using Xunit;
 
 namespace Oci.ObjectstorageService.Transfer
-{ 
+{
     [ExcludeFromCodeCoverage]
     public class ProgressTrackingInputStreamTests : BaseTest
     {